Add files via upload
AI Fitness Trainer

This project is an innovative application of Artificial Intelligence in the field of personal fitness. It leverages advanced computer vision techniques to track and analyze user movements in real-time.

 Features
- Real-time Posture Analysis**: The application uses OpenCV and Mediapipe to track user movements and analyze their exercise form.
- Personalized Feedback**: Based on the analysis, the application provides real-time feedback to help users improve their form and prevent injuries.
- Exercise Recognition**: The application count reps and predict whether the users form is correct or not , making it easier for users to track their progress.
- Interactive Interface**: The application features an easy-to-use interface, making it accessible to everyone, regardless of their tech-savvy.

Future Work
We aim to incorporate more advanced AI models to increase the variety of recognizable exercises and improve the accuracy of our posture analysis. We also plan to develop a feature for personalized workout recommendations based on user's fitness level and goals.

This project is a step towards making personal fitness more accessible and effective through the power of AI.